runner-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
python-dotenv==1.0.0
redis==5.0.1
docker==7.0.0
google-cloud-run==0.10.0
google-cloud-logging==3.8.0
